+++ WebApplication/Views/Employees/Index.cshtml
@@ -1,0 +1,37 @@
+﻿@model IEnumerable<WebApplication.Models.Employee>
+
+<div class="container">
+    <h3>Список сотрудников</h3>
+
+    <table class="table">
+        <thead>
+        <tr>
+            <th>Id</th>
+            <th>Фамилия</th>
+            <th>Имя</th>
+            <th>Отчество</th>
+            <th>Возраст</th>
+            <th></th>
+        </tr>
+        </thead>
+        <tbody>
+        @foreach (var employee in Model)
+        {
+            <tr>
+                <td>@employee.Id</td>
+                <td>@employee.LastName</td>
+                <td>@employee.FirstName</td>
+                <td>@employee.Patronymic</td>
+                <td>@employee.Age</td>
+                <td>
+                    @Html.ActionLink("Карточка сотрудника", "EmployeeInfo", new {id = employee.Id})
+                </td>
+            </tr>
+        }
+        </tbody>
+    </table>
+
+    <hr/>
+
+    <a asp-controller="Home" asp-action="Index">Главная</a>
+</div>
