--- WebApplication/Views/Home/Employees.cshtml
@@ -1,46 +1,0 @@
-﻿@model IEnumerable<WebApplication.Models.Employee>
-<!DOCTYPE html>
-
-<html>
-<head>
-    <meta name="viewport" content="width=device-width" />
-    <title>Employees</title>
-</head>
-<body>
-
-<h3>Список сотрудников</h3>
-
-<table>
-    <thead>
-    <tr>
-        <th>Id</th>
-        <th>Фамилия</th>
-        <th>Имя</th>
-        <th>Отчество</th>
-        <th>Возраст</th>
-        <th></th>
-    </tr>
-    </thead>
-    <tbody>
-    @foreach (var employee in Model)
-    {
-        <tr>
-            <td>@employee.Id</td>
-            <td>@employee.LastName</td>
-            <td>@employee.FirstName</td>
-            <td>@employee.Patronymic</td>
-            <td>@employee.Age</td>
-            <td>
-                @Html.ActionLink("Карточка сотрудника", "EmployeeInfo", new {id = employee.Id})
-            </td>
-        </tr>
-    }
-    </tbody>
-</table>
-
-<hr/>
-
-<a href="~/Home/Index">Главная</a>
-
-</body>
-</html>
