+++ WebApplication/Views/Shared/Partial/_MainMenu.cshtml
@@ -1,0 +1,23 @@
+﻿<div class="mainmenu pull-left">
+    <ul class="nav navbar-nav collapse navbar-collapse">
+        <li><a asp-controller="Home" asp-action="Index" class="active">Home</a></li>
+        <li class="dropdown"><a href="#">Магазин<i class="fa fa-angle-down"></i></a>
+            <ul role="menu" class="sub-menu">
+                <li><a href="shop.html">Каталог</a></li>
+                <li><a href="product-details.html">О товаре</a></li> 
+                <li><a href="checkout.html">Оформить заказ</a></li> 
+                <li><a href="cart.html">Корзина</a></li> 
+                <li><a href="login.html">Вход</a></li> 
+            </ul>
+        </li> 
+        <li class="dropdown"><a href="#">Блог<i class="fa fa-angle-down"></i></a>
+            <ul role="menu" class="sub-menu">
+                <li><a href="blog.html">Список блогов</a></li>
+                <li><a href="blog-single.html">Блог магазина</a></li>
+            </ul>
+        </li> 
+        <li><a href="404.html">404</a></li>
+        <li><a href="contact-us.html">Контакты</a></li>
+        <li><a asp-controller="Employees" asp-action="Index">Сотрудники</a></li>
+    </ul>
+</div>
